requirements.txt
pyinstaller==6.18.0
pillow=12.1.0